remove cycle, switch from symbol to weakmap
